Refactored the Lark Transformer code.
